move rikai.vision to rikai.types.vision
