+++ python/rikai/spark/types/__init__.py
@@ -21,10 +21,8 @@ from pyspark.sql import Row
 from pyspark.sql.types import (
     ArrayType,
     BinaryType,
-    DataType,
     IntegerType,
     ShortType,
-    StringType,
     StructField,
     StructType,
     UserDefinedType,
@@ -34,7 +32,13 @@ from pyspark.sql.types import (
 import rikai
 from rikai.convert import PortableDataType
 from rikai.logging import logger
-from rikai.types.geometry import PointType, Box3dType, Box2dType
+from rikai.spark.types.geometry import Box2dType, Box3dType, PointType
+from rikai.spark.types.vision import (
+    ImageType,
+    LabelType,
+    VideoStreamType,
+    YouTubeVideoType,
+)
 
 __all__ = [
     "ImageType",
@@ -46,41 +50,6 @@ __all__ = [
     "VideoStreamType",
     "YouTubeVideoType",
 ]
-
-
-class ImageType(UserDefinedType):
-    """ImageType defines the Spark UserDefineType for Image type"""
-
-    def __init__(self):
-        super().__init__()
-        self.codec = "png"
-
-    def __repr__(self) -> str:
-        return f"ImageType(codec={self.codec})"
-
-    @classmethod
-    def sqlType(cls) -> StructType:
-        return StructType(fields=[StructField("uri", StringType(), nullable=False)])
-
-    @classmethod
-    def module(cls) -> str:
-        return "rikai.spark.types"
-
-    @classmethod
-    def scalaUDT(cls) -> str:
-        return "org.apache.spark.sql.rikai.ImageType"
-
-    def serialize(self, obj: "Image"):
-        """Serialize an Image to a Spark Row?"""
-        return (obj.uri,)
-
-    def deserialize(self, datum) -> "Image":
-        from rikai.vision import Image  # pylint: disable=import-outside-toplevel
-
-        return Image(datum[0])
-
-    def simpleString(self) -> str:
-        return "ImageType"
 
 
 class NDArrayType(UserDefinedType):
@@ -141,108 +110,3 @@ class NDArrayType(UserDefinedType):
 
     def simpleString(self) -> str:
         return "ndarray"
-
-
-class LabelType(UserDefinedType):
-    """Label type"""
-
-    def __repr__(self) -> str:
-        return "LabelType"
-
-    @classmethod
-    def sqlType(cls) -> DataType:
-        return StringType()
-
-    @classmethod
-    def module(cls) -> str:
-        return "rikai.spark.types"
-
-    @classmethod
-    def scalaUDT(cls) -> str:
-        return "org.apache.spark.sql.rikai.LabelType"
-
-    def serialize(self, obj: "Label"):
-        """Serialize a label into Spark String"""
-        return obj.label
-
-    def deserialize(self, datum: Row) -> "Label":
-        from rikai.vision import Label
-
-        return Label(datum)
-
-    def simpleString(self) -> str:
-        return "label"
-
-
-class VideoStreamType(UserDefinedType):
-    """VideoStreamType defines the Spark UserDefineType for
-    a given video stream
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.codec = "mp4"  # TODO allow generic code
-
-    def __repr__(self) -> str:
-        return f"VideoType(codec={self.codec})"
-
-    @classmethod
-    def sqlType(cls) -> StructType:
-        return StructType(fields=[StructField("uri", StringType(), nullable=False)])
-
-    @classmethod
-    def module(cls) -> str:
-        return "rikai.spark.types"
-
-    @classmethod
-    def scalaUDT(cls) -> str:
-        return "org.apache.spark.sql.rikai.VideoStreamType"
-
-    def serialize(self, obj: "VideoStream"):
-        """Serialize a VideoStream to a Spark Row"""
-        return (obj.uri,)
-
-    def deserialize(self, datum) -> "VideoStream":
-        from rikai.vision import VideoStream  # pylint: disable=import-outside-toplevel
-
-        return VideoStream(datum[0])
-
-    def simpleString(self) -> str:
-        return "VideoStreamType"
-
-
-class YouTubeVideoType(UserDefinedType):
-    """YouTubeVideoType defines the Spark UserDefineType for
-    a piece of YouTube video content (i.e., corresponds to a given
-    youtube id but can have multiple streams)
-    """
-
-    def __init__(self):
-        super().__init__()
-
-    def __repr__(self) -> str:
-        return "YouTubeVideoType"
-
-    @classmethod
-    def sqlType(cls) -> StructType:
-        return StructType(fields=[StructField("vid", StringType(), nullable=False)])
-
-    @classmethod
-    def module(cls) -> str:
-        return "rikai.spark.types"
-
-    @classmethod
-    def scalaUDT(cls) -> str:
-        return "org.apache.spark.sql.rikai.YouTubeVideoType"
-
-    def serialize(self, obj: "YouTubeVideo"):
-        """Serialize an Image to a Spark Row?"""
-        return (obj.vid,)
-
-    def deserialize(self, datum) -> "YouTubeVideo":
-        from rikai.vision import YouTubeVideo  # pylint: disable=import-outside-toplevel
-
-        return YouTubeVideo(datum[0])
-
-    def simpleString(self) -> str:
-        return "YouTubeVideoType"
