+++ src/main/scala/org/apache/spark/sql/rikai/Video.scala
@@ -36,10 +36,10 @@ class VideoStreamType extends UserDefinedType[VideoStream] {
     StructType(
       Seq(
         StructField("uri", StringType, false)
-        )
       )
+    )
 
-  override def pyUDT: String = "rikai.spark.types.VideoStreamType"
+  override def pyUDT: String = "rikai.spark.types.vision.VideoStreamType"
 
   override def serialize(obj: VideoStream): Any = {
     val row = new GenericInternalRow(1);
@@ -77,10 +77,10 @@ class YouTubeVideoType extends UserDefinedType[YouTubeVideo] {
     StructType(
       Seq(
         StructField("vid", StringType, false)
-        )
       )
+    )
 
-  override def pyUDT: String = "rikai.spark.types.YouTubeVideoType"
+  override def pyUDT: String = "rikai.spark.types.vision.YouTubeVideoType"
 
   override def serialize(obj: YouTubeVideo): Any = {
     val row = new GenericInternalRow(1);
